Feature: Exact Change Only
Several scope limitations were made to time-box the implementation of
this feature to complete the kata in a manner that indicates my thought
processes without implementing a full algorithm to determine whether the
precise collection of coins in the vending machine's coin changer can be
used to return a specific amount.  Where very small numbers of nickels
and dimes are present, some amounts cannot be successfully returned to
the coin return even though the total value of the coins exceeds that
amount.  Comments have been placed throughout the code to indicate where
assumptions have been made, and the coin changer contents have been
initialized to prevent the special-case conditions from occurring in the
acceptance tests.  Error-handling related to these conditions can also
be omitted for the purposes of considering this feature to be complete
enough for the kata exercise.